double edited

+++ com.gmail.lesson2.1/src/com/gmail/lessoncom/gmail/lesson/Main.java
@@ -8,11 +8,11 @@ public class Main {
 		Scanner sc = new Scanner(System.in);
 
 		System.out.println("введите число 1");
-		int a = sc.nextInt();
+		double a = sc.nextInt();
 		System.out.println("введите число 2");
-		int b = sc.nextInt();
+		double b = sc.nextInt();
 
-		int c = (a + b) / 2;
+		double c = (a + b) / 2;
 
 		System.out.println("среднее число = " + c);
